#1331: extract submit_cos_batch per-variant handlers
Extract the 221-LOC submit_cos_batch() match dispatcher in
userspace-dp/src/afxdp/cos/queue_service/mod.rs into two
flat-sibling per-variant handler modules:

  queue_service/
    submit_local.rs     pub(super) #[inline] fn submit_local
                         (145 LOC, includes the moved private
                          restore_cos_local_items helper)
    submit_prepared.rs  pub(super) #[inline] fn submit_prepared
                         (140 LOC, includes the moved private
                          restore_cos_prepared_items helper)

submit_cos_batch in mod.rs is now a thin 43-LOC match shim that
destructures CoSBatch::{Local,Prepared} and forwards the fields
to the matching handler. Both handlers are #[inline] so LLVM
inlines them across the cgu boundary at -O3, leaving the
generated machine code equivalent to the pre-refactor monolith.

Pure code motion — match-arm bodies copied verbatim. No control-flow
change, no allocation change (the per-batch [(u16,u64); TX_BATCH_SIZE]
stack sidecar stays inside each handler), no visibility bumps
(child modules access private parent items via super::). The
#1229 v7 flow-fair sidecar logic, #760 instrumentation order,
#710 tx_submit_error_drops accounting, and error-path restore
semantics are all preserved bit-for-bit.

restore_cos_{local,prepared}_items moved from mod.rs into the
owning child (each had exactly one caller). The companion
restore_cos_{local,prepared}_items_inner helpers stay in
tx_completion since they have other consumers.

Plan + adversarial review

Plan doc: docs/pr/1331-submit-cos-batch-per-variant/plan.md

  Round-1 verdicts:
    AGY    PLAN-NEEDS-MAJOR — drop submit/ subdir, use flat siblings
    Codex  PLAN-NEEDS-MAJOR — Local sidecar prefix invariant is FALSE
                              because transmit_batch (transmit.rs:96-109)
                              can drop mirror_clone items via continue,
                              breaking the head-prefix assumption. Pre-
                              existing accounting defect; refactor
                              preserves it verbatim.
    Gemini PLAN-NEEDS-MINOR — drop submit/ subdir
    Claude PLAN-READY-pending-revision (hostile-verified #1561 race
           surface unchanged: null deref is upstream Arc publish,
           submit_cos_batch consumes by value past the resolution
           point)

  v2 changes folded in:
    - Flat sibling layout queue_service/submit_local.rs +
      queue_service/submit_prepared.rs (matches drain.rs /
      service.rs / tests.rs pattern).
    - No visibility bumps (Rust child modules see private parent
      items via super::).
    - mirror_clone sidecar prefix-attribution defect documented as
      an explicit carry-over hazard with a deferred follow-up.

  Round-2 verdicts on plan v2 (commit 92eeb1bf):
    AGY    PLAN-READY
    Codex  PLAN-NEEDS-MINOR — doc-only stale v1 prose at plan.md
                              lines 77-79 / 110-112 / 449-450 still
                              referenced the submit/ subdir
    Gemini PLAN-READY
    Claude PLAN-READY

  Doc cleanup pushed; plan v2 is PLAN-READY across all four
  reviewers.

Verification

  - cargo build clean (122 pre-existing warnings, 0 errors).
  - cargo test --release: 1503 tests pass.
    One pre-existing master failure unrelated to this refactor —
    snat_contract_documents_current_fail_closed_runtime asserts
    that docs/userspace-dataplane-gaps.md contains the word
    "fail-closed". That file at master b92198f8 does not contain
    the word; the failure reproduces on stock master with no
    changes from this branch.
  - 5/5 flake check on afxdp::cos::queue_service::* — 49 tests pass
    each iteration.
  - Go suite passes.
  - mod.rs: 1536 → 1325 LOC (−211). submit_cos_batch: 221 → 43 LOC.

#1561 (open) race compatibility

  The first-snapshot drop_in_place<CoSBatch> null-deref crash is
  upstream of submit_cos_batch in the coordinator's snapshot
  publish path. submit_cos_batch consumes CoSBatch by value after
  the upstream Arc has been resolved; the refactor adds at most
  one extra memcpy of the destructured VecDeque header bytes
  caller→callee, which reads the same memory the inline match arm
  reads. CoSBatch enum shape, Drop, and publish path are all
  unchanged. No new ArcSwap or shared-state read introduced. Race
  observability is identical before and after.

Out of scope (deferred, tracked separately)

  - Mirror-clone sidecar prefix-attribution defect on the Local
    path. Pre-existing accounting bug. Refactor preserves the
    buggy-but-stable behavior verbatim. Fix requires reshaping
    transmit_batch to surface the post-skip submit order to the
    sidecar build.
  - #1207 queue_service/service.rs consolidation.
  - #1206 CoSQueueRuntime split.

Closes #1331
