+++ docs/pr/1331-submit-cos-batch-per-variant/reviewer-ids.md
@@ -6,4 +6,5 @@ session state is lost (per `feedback_codex_session_loss_continuation`).
 
 | Round | Codex task-id | AGY task-id | Gemini task-id | Notes |
 |---|---|---|---|---|
-| PLAN v1 | task-mpn2uc3f-bfc18j | review-mpn2vgin-13r8jz | task-mpn2v4xq-m9fqna | dispatched 20:18 UTC 2026-05-26 |
+| PLAN v1 | task-mpn2uc3f-bfc18j | review-mpn2vgin-13r8jz | task-mpn2v4xq-m9fqna | dispatched 20:18 UTC; Codex/AGY MAJOR, Gemini MINOR |
+| PLAN v2 | task-mpn32ftd-8uza6f | review-mpn32taj-4iawxl | task-mpn32wlj-v1nrhu | dispatched 20:24 UTC 2026-05-26 |
